No smoking (#263)
In future events, we should mention the no-smoking rule right at the beginning. The sign at the c-base bar is too easily overlooked.

One might argue that this is more a rule of the venue than a rule of the event, but that wouldn't be accurate: c-base is smoking-permitted by default, and BHNT is one of the events who opted to be non-smoking.

During initial presentation, one might mention that smoking is permitted outside, but I didn't want to clutter the rules page with that level of detail.